% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workdocs_operations.R
\name{workdocs_activate_user}
\alias{workdocs_activate_user}
\title{Activates the specified user}
\usage{
workdocs_activate_user(UserId, AuthenticationToken)
}
\arguments{
\item{UserId}{[required] The ID of the user.}

\item{AuthenticationToken}{Amazon WorkDocs authentication token. Do not set this field when using
administrative API actions, as in accessing the API using AWS
credentials.}
}
\description{
Activates the specified user. Only active users can access Amazon
WorkDocs.
}
\section{Request syntax}{
\preformatted{workdocs$activate_user(
  UserId = "string",
  AuthenticationToken = "string"
)
}
}

\keyword{internal}
