% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workdocs_operations.R
\name{workdocs_create_folder}
\alias{workdocs_create_folder}
\title{Creates a folder with the specified name and parent folder}
\usage{
workdocs_create_folder(AuthenticationToken, Name, ParentFolderId)
}
\arguments{
\item{AuthenticationToken}{Amazon WorkDocs authentication token. Do not set this field when using
administrative API actions, as in accessing the API using AWS
credentials.}

\item{Name}{The name of the new folder.}

\item{ParentFolderId}{[required] The ID of the parent folder.}
}
\description{
Creates a folder with the specified name and parent folder.
}
\section{Request syntax}{
\preformatted{workdocs$create_folder(
  AuthenticationToken = "string",
  Name = "string",
  ParentFolderId = "string"
)
}
}

\keyword{internal}
